Add GoogleInputTools swift package (#1)
This is a client package that talks to Google Input Service.

The service response is a deeply nested json array of mixed types. It makes
parsing a bit difficult. There are a few good tests to cover the different
cases I could find.

Wire it up to the keyboard with basic functionality where the word suggestions
would appear in the autocomplete toolbar. It is basic and buggy.